+++ swing/framework-model/src/main/java/is/codion/swing/framework/model/SwingEntityEditModel.java
@@ -46,7 +46,8 @@ public class SwingEntityEditModel extends DefaultEntityEditModel {
 	 */
 	public SwingEntityEditModel(EntityType entityType, EntityConnectionProvider connectionProvider, SwingComponentModels componentModels) {
 		super(new SwingEntityEditor(entityType, connectionProvider, componentModels));
-		afterInsertUpdateOrDelete().addListener(editor().comboBoxModels()::refreshColumnComboBoxModels);
+		afterInsertUpdateOrDelete().addListener(() -> editor().comboBoxModels().column().values()
+						.forEach(comboBoxModel -> comboBoxModel.items().refresh()));
 	}
 
 	@Override